# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /store-types:
    get:
      operationId: listStoreTypes
      tags: [StoreTypes]
      summary: List curated store types (public)
      responses:
        "200":
          description: Store types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreType"
    post:
      operationId: createStoreType
      tags: [StoreTypes]
      summary: Add a store type (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreTypeInput"
      responses:
        "201":
          description: Created store type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreType"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Duplicate type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /store-types/{id}:
    patch:
      operationId: updateStoreType
      tags: [StoreTypes]
      summary: Update a store type (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreTypeUpdate"
      responses:
        "200":
          description: Updated store type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreType"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteStoreType
      tags: [StoreTypes]
      summary: Delete a store type (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves object entities uploaded via presigned URLs.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
...
    StoreType:
      type: object
      required: [id, name, sortOrder, showRecipes, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        showRecipes:
          type: boolean
          description: Whether stores of this type show the "شنو نطبخ اليوم؟" recipe helper.
        createdAt:
          type: string
          format: date-time
    StoreTypeInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        showRecipes:
          type: boolean
    StoreTypeUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        showRecipes:
          type: boolean
...